Created synchapi.py Moved functions from kernel32.py to syncapi.py if they are declader in synchapi.h
