+++ qiling/os/windows/dlls/kernel32.py
@@ -1208,16 +1208,9 @@ def hook_GlobalUnlock(ql, address, params):
 def hook_GlobalAlloc(ql, address, params):
     return ql.heap.mem_alloc(params['dwBytes'])
 
-#BOOL InitializeCriticalSectionAndSpinCount(
-#  LPCRITICAL_SECTION lpCriticalSection,
-#  DWORD              dwSpinCount
-#);
-@winapi(cc=STDCALL, params={
-    "lpCriticalSection": POINTER,
-    "dwSpinCount": UINT
-})
-def hook_InitializeCriticalSectionAndSpinCount(ql, address, params):
-    return 1
+
+
+
 
 #DWORD TlsAlloc();
 @winapi(cc=STDCALL, params={})
@@ -1369,23 +1362,6 @@ def hook_SetLastError(ql, address, params):
 def hook_IsValidCodePage(ql, address, params):
     return 1
 
-#void EnterCriticalSection(
-#  LPCRITICAL_SECTION lpCriticalSection
-#);
-@winapi(cc=STDCALL, params={
-    "lpCriticalSection": POINTER
-})
-def hook_EnterCriticalSection(ql, address, params):
-    return 0
-
-#void LeaveCriticalSection(
-#  LPCRITICAL_SECTION lpCriticalSection
-#);
-@winapi(cc=STDCALL, params={
-    "lpCriticalSection": POINTER
-})
-def hook_LeaveCriticalSection(ql, address, params):
-    return 0
 
 #int MultiByteToWideChar(
 #  UINT                              CodePage,
